# docs/swagger.yaml
basePath: /
definitions:
  errors.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      status:
        type: integer
    type: object
  handlers.addTimeSlotRequest:
    properties:
      day_of_week:
        description: time.Weekday
        type: integer
      hour:
        type: integer
      minute:
        type: integer
    type: object
  handlers.addUserRequest:
    properties:
      email:
        type: string
    type: object
...
  handlers.createProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  project.Project:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      idea_queue:
        items:
          type: string
        type: array
      name:
        type: string
      post_queue:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
...
  project.UserPlatformInfo:
    properties:
      authTTL:
        type: string
      isAuthenticated:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for OpenCM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OpenCM API
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: List all projects that the user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.Project'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: List all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project with the given name and description
      parameters:
      - description: Project creation request
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/handlers.createProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/project.Project'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "409":
          description: Project already exists
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Create a new project
      tags:
      - projects
  /projects/{project_id}:
    delete:
      consumes:
      - application/json
      description: Delete a project with the given ID
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
          schema:
            type: string
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Project not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Delete a project
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Get a project by its ID
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.Project'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Project not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Get a project
      tags:
      - projects
    patch:
      consumes:
      - application/json
      description: Update a project with the given ID
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Project update request
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/handlers.createProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.Project'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Project not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Update a project
      tags:
      - projects
  /projects/{project_id}/add-role/{user_id}/{role_id}:
    post:
      consumes:
      - application/json
      description: Add a role to a user in a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Role ID
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
          schema:
            type: string
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Project not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Add a role to a user
      tags:
      - projects
  /projects/{project_id}/add-time-slot:
    patch:
      consumes:
      - application/json
      description: Add a time slot to a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Time slot request
        in: body
        name: time_slot
        required: true
        schema:
          $ref: '#/definitions/handlers.addTimeSlotRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No content
          schema:
            type: string
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Project not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Add a time slot to a project
      tags:
      - projects
  /projects/{project_id}/add-user:
    post:
      consumes:
      - application/json
      description: Add a user to a project by their email
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: User addition request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.addUserRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No content
          schema:
            type: string
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Project not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Add a user to a project
      tags:
      - projects
  /projects/{project_id}/default-user-platform-info/{platform_id}:
    get:
      consumes:
      - application/json
      description: Get the default user platform info for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Platform ID
        in: path
        name: platform_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.UserPlatformInfo'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Project not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Get the default user platform info
      tags:
      - projects
  /projects/{project_id}/default-user/{user_id}:
    patch:
      consumes:
      - application/json
      description: Set the default user for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
          schema:
            type: string
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Project not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Set the default user for a project
      tags:
      - projects
...
schemes:
- http
- https
swagger: "2.0"